+++ parser/pas.go
@@ -1,10 +1,9 @@
-package pas
+package parser
 
 import (
 	"github.com/akm/pas/ast"
-	"github.com/akm/pas/parser"
 )
 
 func ParseString(code string) (*ast.File, error) {
-	return parser.ParseString(code)
+	return new([]rune(code)).parseFile()
 }
